Refactor sound playback functionality in AdminTable and Table components to toggle between Play and Pause icons
